boat_article_matcher/requirements.txt
# Python >=3.8 required
asyncpg>=0.28.0     # Асинхронная работа с PostgreSQL
pandas>=2.0.0       # Обработка CSV файлов
python-dotenv>=1.0.0 # Загрузка переменных окружения из .env
loguru>=0.7.0       # Улучшенное логирование с ротацией